+++ example/lib/model/method.dart
@@ -29,6 +29,7 @@ class Method {
   static List<String> METHOD_FOR_CARD = ['card'];
   static List<String> METHOD_FOR_PHONE = ['phone'];
   static List<String> METHOD_FOR_VBANK = ['vbank'];
+  static List<String> METHOD_FOR_TRANS = ['trans'];
 
   static String getLabel(String method) {
     switch (method) {
@@ -63,6 +64,9 @@ class Method {
         return 'phone';
       case 'settle':
         return 'vbank';
+      case 'chai':
+      case 'payple':
+        return 'trans';
       default:
         return 'card';
     }
@@ -80,6 +84,7 @@ class Method {
       case 'paypal':
       case 'payco': 
       case 'smilepay':
+      case 'alipay':
         return METHOD_FOR_CARD;
       case 'uplus':
         return METHODS_FOR_UPLUS;
@@ -89,6 +94,9 @@ class Method {
         return METHODS_FOR_MOBILIANS;
       case 'settle':
         return METHOD_FOR_VBANK;
+      case 'chai':
+      case 'payple':
+        return METHOD_FOR_TRANS;
       default:
         return METHODS;
     }