+++ CHANGELOG.md
@@ -1,4 +1,29 @@
-## [0.9.4](https://github.com/iamport/iamport_flutter/tree/master)
+## [0.9.12](https://github.com/iamport/iamport_flutter/tree/master)
+- [안드로이드] KG이니시스 - 실시간계좌이체시 국민리브, NH앱캐시, NG상상뱅크, BNK경남은행 앱의 링크가 누락된 부분을 추가하였습니다.
+- uni_links 패키지의 버전을 0.4.0으로 올렸습니다.
+
+## [0.9.11](https://github.com/iamport/iamport_flutter/tree/v0.9.11)
+- 스마일페이, 차이페이, 페이플, 알리페이 예제 코드를 추가하였습니다.
+- 다날 - 본인인증 방식을 리디렉션 방식으로 변경하였습니다.
+
+## [0.9.10](https://github.com/iamport/iamport_flutter/tree/v0.9.10)
+- iamport javascript sdk 버전을 v1.1.8로 올렸습니다.
+- 안드로이드6에서 결제/본인인증 창 미렌더링 이슈를 해결하였습니다.
+
+## [0.9.9](https://github.com/iamport/iamport_flutter/tree/v.0.9.9)
+- 농협 올원페이 앱 스킴 값 오타를 수정하였습니다.
+- build시 필요하지 않은 asset을 제거하기 위해 pubspec.yaml 파일을 수정하였습니다.
+
+## [0.9.8](https://github.com/iamport/iamport_flutter/tree/v0.9.8)
+- 할부개월수 설정 파라미터 오타를 고쳤습니다.
+
+## [0.9.6](https://github.com/iamport/iamport_flutter/tree/v0.9.6)
+- niceMobileV2를 true로 기본으로 적용하고 이와 관련된 실시간 계좌이체 대비 코드를 적용하였습니다.
+
+## [0.9.5](https://github.com/iamport/iamport_flutter/tree/v0.9.5)
+- 이니시스, 나이스 그리고 다날 일반결제시 제공기간 표기를 위한 파라미터(period)를 추가하였습니다.
+
+## [0.9.4](https://github.com/iamport/iamport_flutter/tree/v0.9.4)
 - 안드로이드에서 NH농협카드 일반결제시, intent uri를 파싱할때 발생하는 exception을 처리하였습니다.
 
 ## [0.9.3](https://github.com/iamport/iamport_flutter/tree/v0.9.3)